RIA-7412: moved handler class to presubmit package
