+++ src/test/java/uk/gov/hmcts/reform/iacaseapi/domain/handlers/presubmit/SpokenLanguageForAppellantCaseFlagsHandlerTest.java
@@ -17,7 +17,6 @@ import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.Callback;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.callback.PreSubmitCallbackStage;
 import uk.gov.hmcts.reform.iacaseapi.domain.entities.ccd.field.YesOrNo;
-import uk.gov.hmcts.reform.iacaseapi.domain.handlers.SpokenLanguageForAppellantCaseFlagsHandler;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
